Allow instruction opcodes to be multi-byte, like literal numbers (fix #14)
Encode them in the same way as numbers.

Overhaul the debug and auxiliary APIs for assembling and disassembling code
to deal with this.

Refer to “numbers” rather than “literals” for instructions encoding literal
numbers, and to “actions” for other instructions.

Rather nicely, this is backwards-compatible: the one-byte opcodes still
encode as the same single byte!

+++ src/public.h
@@ -44,7 +44,7 @@ typedef uint64_t DUWORD;
 #endif
 
 extern UWORD PC;
-extern BYTE I;
+extern WORD I;
 extern UWORD MEMORY;
 extern UWORD SP, RP;
 extern UWORD S0, R0;
